requirements.txt
asgiref==3.8.1
boto3==1.38.33
botocore==1.38.33
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
Django==5.2
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
enum-compat==0.0.3
Faker==37.1.0
idna==3.10
inflection==0.5.1
intuit-oauth==1.2.6
jmespath==1.0.1
load-dotenv==0.1.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-quickbooks==0.9.12
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
